improves image search

+++ src/main/java/de/toomuchcoffee/figurearchive/repository/FigureRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FigureRepository extends JpaRepository<Figure, Integer> {
+public interface FigureRepository extends JpaRepository<Figure, Long> {
     List<Figure> findByVerbatimStartsWithIgnoreCase(String verbatim);
     List<Figure> findByProductLine(ProductLine productLine);
     List<Figure> findByVerbatimStartsWithIgnoreCaseAndProductLine(String verbatim, ProductLine productLine);
